refactor: remove pass-through facade re-exports (#462)
Eliminate 4 redundant pass-through re-export modules whose consumers
can import directly from the source:

- src/utils/dom-facades.js (39 consumers -> import from dom.js)
- src/utils/flow-dom.js (6 consumers -> import from dom.js)
- src/utils/chevron-row.js (4 consumers -> import from dom.js)
- main/managers.js (1 consumer -> inlined into manager-init.js)

The 4 subsystem facades (file-tree-subsystem, file-viewer-subsystem,
flow-view-subsystem, terminal-subsystem) are kept: they aggregate
multiple source modules and serve 2-3 consumers each, providing
genuine coupling reduction.

Closes #462
