+++ src/components/flow-modal.js
@@ -1,5 +1,5 @@
 import { generateId } from '../utils/id.js';
-import { _el, createActionButton } from '../utils/flow-dom.js';
+import { _el, createActionButton } from '../utils/dom.js';
 import { createModalOverlay } from '../utils/dom-dialogs.js';
 import {
   SCHEDULE_LABELS, DAY_NAMES, WEEKDAY_INDICES, INTERVAL_HOURS,
