+++ src/utils/file-tree-drop.js
@@ -4,7 +4,7 @@
  */
 
 import { emitFileOpen } from './workspace-events.js';
-import { _el } from './dom-facades.js';
+import { _el } from './dom.js';
 import { setupInlineInput, startInlineRename } from './form-helpers.js';
 import { setupDropZone as _setupDropZone } from './drop-zone-helpers.js';
 import { INPUT_BLUR_DELAY, computeIndent, getBaseName } from './file-tree-helpers.js';
