+++ src/utils/file-tree-section-dom.js
@@ -8,8 +8,8 @@
  * @typedef {{ setupDropZone: (el: HTMLElement, targetDir: string|(() => string|null)) => void, promptNewEntry: (dirPath: string, cEl: HTMLElement, depth: number, eDirs: Set<string>, type: string) => void, promptRename: (path: string, nameEl: HTMLElement) => void, refreshSection: (cwd: string) => void, contextMenuApi: unknown }} SectionDOMCallbacks
  */
 
-import { _el } from './dom-facades.js';
-import { buildChevronRow } from './chevron-row.js';
+import { _el } from './dom.js';
+import { buildChevronRow } from './dom.js';
 import { attachContextMenu } from './context-menu.js';
 import { emitWorkspaceCreateWorktree, emitWorkspaceOpenPr } from './workspace-events.js';
 import {
