+++ src/utils/file-viewer-tabs.js
@@ -3,7 +3,7 @@
  * Extracted from file-viewer.js to reduce component size.
  */
 
-import { _el, renderList } from './dom-facades.js';
+import { _el, renderList } from './dom.js';
 import { attachContextMenu } from './context-menu.js';
 import { createTabElement } from './tab-renderer.js';
