+++ src/utils/terminal-subsystem.js
@@ -3,7 +3,7 @@
  * and board-view components to access terminal infrastructure.
  *
  * Covers the split-layout, serialization, node-building, drag-drop
- * indicator, terminal-split, terminal-events, dom-facades, and
+ * indicator, terminal-split, terminal-events, dom, and
  * terminal-factory APIs (8 source modules).
  *
  * Reduces the import surface of terminal-panel.js and board-view.js.
@@ -44,8 +44,8 @@ export {
   onTerminalCreated, onTerminalRemoved, onTerminalExited,
 } from './terminal-events.js';
 
-// ── dom-facades (board-view) ────────────────────────────────────────
-export { _el, renderButtonBar, buildDomainButtonBar, renderList } from './dom-facades.js';
+// ── dom (board-view) ────────────────────────────────────────────────
+export { _el, renderButtonBar, buildDomainButtonBar, renderList } from './dom.js';
 
 // ── terminal-factory (board-view) ───────────────────────────────────
 export {
