+++ src/utils/worktree-flow.js
@@ -9,7 +9,7 @@
 
 import { showWorktreeDialog } from './worktree-dialog.js';
 import { showConfirmDialog } from './dom-dialogs.js';
-import { _el } from './dom-facades.js';
+import { _el } from './dom.js';
 import { gitFlowStep } from './git-flow-helpers.js';
 
 /**
